drivers: ethernet: nxp: move into own dir
move nxp ethernet drivers into own directory.
