+++ drivers/ethernet/CMakeLists.txt
@@ -6,12 +6,6 @@ zephyr_library_include_directories(${ZEPHYR_BASE}/subsys/net/l2)
 
 add_subdirectory_ifdef(CONFIG_MDIO mdio)
 
-if(CONFIG_ETH_NXP_S32_NETC)
-  zephyr_library_sources(eth_nxp_s32_netc.c)
-  zephyr_library_sources_ifdef(CONFIG_DT_HAS_NXP_S32_NETC_PSI_ENABLED eth_nxp_s32_netc_psi.c)
-  zephyr_library_sources_ifdef(CONFIG_DT_HAS_NXP_S32_NETC_VSI_ENABLED eth_nxp_s32_netc_vsi.c)
-endif()
-
 if(CONFIG_ETH_STM32_HAL)
   if(CONFIG_ETH_STM32_HAL_RX_DESC_COUNT)
     zephyr_compile_definitions(ETH_RX_DESC_CNT=${CONFIG_ETH_STM32_HAL_RX_DESC_COUNT})
@@ -34,8 +28,7 @@ endif()
 
 add_subdirectory(dsa)
 add_subdirectory(phy)
-add_subdirectory(eth_nxp_enet_qos)
-add_subdirectory(nxp_imx_netc)
+add_subdirectory(nxp)
 add_subdirectory(dwc_mac)
 add_subdirectory(dwc_xgmac)
 add_subdirectory(intel)
@@ -59,8 +52,6 @@ zephyr_library_sources_ifdef(CONFIG_ETH_LAN9250 eth_lan9250.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_LITEX_LITEETH eth_litex_liteeth.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_MCHP_GMAC_G1 eth_mchp_gmac_g1.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_NUMAKER eth_numaker.c)
-zephyr_library_sources_ifdef(CONFIG_ETH_NXP_ENET eth_nxp_enet.c)
-zephyr_library_sources_ifdef(CONFIG_ETH_NXP_S32_GMAC eth_nxp_s32_gmac.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_RENESAS_RA eth_renesas_ra.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_RENESAS_RA_RMAC eth_renesas_ra_rmac.c)
 zephyr_library_sources_ifdef(CONFIG_ETH_SAM_GMAC eth_sam_gmac.c)
