+++ drivers/ethernet/Kconfig
@@ -66,9 +66,6 @@ source "drivers/ethernet/Kconfig.litex"
 source "drivers/ethernet/Kconfig.mchp"
 source "drivers/ethernet/Kconfig.native_tap"
 source "drivers/ethernet/Kconfig.numaker"
-source "drivers/ethernet/Kconfig.nxp_enet"
-source "drivers/ethernet/Kconfig.nxp_s32_gmac"
-source "drivers/ethernet/Kconfig.nxp_s32_netc"
 source "drivers/ethernet/Kconfig.renesas_ra"
 source "drivers/ethernet/Kconfig.renesas_ra_rmac"
 source "drivers/ethernet/Kconfig.sam_gmac"
@@ -90,9 +87,8 @@ source "drivers/ethernet/Kconfig.xmc4xxx"
 source "drivers/ethernet/dsa/Kconfig"
 source "drivers/ethernet/dwc_mac/Kconfig"
 source "drivers/ethernet/dwc_xgmac/Kconfig"
-source "drivers/ethernet/eth_nxp_enet_qos/Kconfig"
 source "drivers/ethernet/intel/Kconfig"
-source "drivers/ethernet/nxp_imx_netc/Kconfig"
+source "drivers/ethernet/nxp/Kconfig"
 source "drivers/ethernet/offload/Kconfig"
 source "drivers/ethernet/phy/Kconfig"
 # zephyr-keep-sorted-stop
