+++ drivers/ethernet/nxp/Kconfig
@@ -1,0 +1,10 @@
+# SPDX-FileCopyrightText: Copyright The Zephyr Project Contributors
+# SPDX-License-Identifier: Apache-2.0
+
+# zephyr-keep-sorted-start
+rsource "Kconfig.nxp_enet"
+rsource "Kconfig.nxp_s32_gmac"
+rsource "Kconfig.nxp_s32_netc"
+rsource "eth_nxp_enet_qos/Kconfig"
+rsource "nxp_imx_netc/Kconfig"
+# zephyr-keep-sorted-stop
