+++ drivers/ethernet/nxp/eth_nxp_s32_netc.c
@@ -23,7 +23,7 @@ LOG_MODULE_REGISTER(nxp_s32_eth);
 #include <Netc_Eth_Ip_Irq.h>
 #include <Netc_EthSwt_Ip.h>
 
-#include "eth.h"
+#include "../eth.h"
 #include "eth_nxp_s32_netc_priv.h"
 
 /* Global MAC filter hash table required for the baremetal driver */
