+++ drivers/ethernet/nxp/nxp_imx_netc/eth_nxp_imx_netc.c
@@ -21,7 +21,7 @@ LOG_MODULE_REGISTER(nxp_imx_eth);
 #ifdef CONFIG_GIC_V3_ITS
 #include <zephyr/drivers/interrupt_controller/gicv3_its.h>
 #endif
-#include "../eth.h"
+#include "../../eth.h"
 #include "eth_nxp_imx_netc_priv.h"
 
 #ifdef CONFIG_ETH_NXP_IMX_MSGINTR
